feat: CommentLayout created  and cleaned up [slug]
