+++ src/hooks/useSubmitComment.ts
@@ -112,7 +112,6 @@ export function useSubmitComment() {
             addComment.mutate(
                 {
                     content,
-                    commenterId,
                     postSlug,
                 },
                 {
